feat(auth): ✨ Refreshes user session on update
Adds functionality to refresh the local user session when user metadata is updated on the backend.

This ensures that changes like `idPersona` are reflected in the client-side session, especially after account creation or migration.
It introduces a `reloadUser` method to fetch the latest user data from Supabase and update the cached session.
Also implements migration by document number.